mainly worked on write_to_database file and made some small changes to other files
write_to_database creates a db and a single table in a local mysql
server.  since I didn't solve the problem of multiple locations I
altered the other files in this commit by removing the listings with
multiple states as their locations.  It brought the total number of
listings down from 8038 to 7710.